requirements.txt
alkana==0.0.3
config==0.5.1
deepl==1.14.0
emoji==1.7.0
googletrans==4.0.0rc1
keyboard==0.13.5
mecab==0.996.3
openai==0.27.4
pandas==1.5.3
pyaudio==0.2.13
pytchat==0.5.5
python-dotenv==1.0.0
Requests==2.28.2
sounddevice==0.4.6
soundfile==0.12.1
SpeechRecognition==3.9.0
torch==1.13.1
unidic==1.1.0
